+++ src/main/java/domain/player/dealer/DealerResult.java
@@ -1,4 +1,4 @@
-package domain.player;
+package domain.player.dealer;
 
 public enum DealerResult {
