+++ src/test/java/domain/player/participant/ParticipantTest.java
@@ -1,8 +1,11 @@
-package domain.player;
+package domain.player.participant;
 
 import domain.area.CardArea;
 import domain.card.Card;
 import domain.card.CardShape;
+import domain.player.Name;
+import domain.player.participant.Participant;
+import domain.player.participant.State;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator;
