+++ src/test/java/domain/player/participant/StateTest.java
@@ -1,5 +1,6 @@
-package domain.player;
+package domain.player.participant;
 
+import domain.player.participant.State;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
